feat: improve performance and add new props
BREAKING CHANGE: items are optimized by reducing shared values in the
parent cells and reducing the animated styles. ReorderableListItem
animations have moved to the cell itself and are attached to the active
cell only. A new prop cellAnimations and onDragStart were added to allow
customizing opacity and scale, while ReorderableListItem is removed from
the api.

Closes #18